First compilation with Dyn4J. No graphics or movement
Refactored and commented out enough to allow compilation and run without errors.
Now need to add each bit back to work with Dyn4J

+++ src/main/java/com/ewan/dunjeon/generation/GeneratorsMisc.java
@@ -146,8 +146,8 @@ public class GeneratorsMisc {
             return String.format("%f, %f -> %f, %f", x1, y1, x2, y2);
         }
 
-        public Line2D.Float getLine2D(){
-            return new Line2D.Float(x1, y1, x2, y2);
+        public Line2D.Double getLine2D(){
+            return new Line2D.Double(x1, y1, x2, y2);
         }
 
         public int getX1() {
